% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_survey.R
\name{compare_survey}
\alias{compare_survey}
\title{Compares survey sheets from two xlsform files}
\usage{
compare_survey(form1, form2, language)
}
\arguments{
\item{form1}{Filepath of first form as text or referring to object with location.}

\item{form2}{Filepath of second form as text or referring to object with location.}

\item{language}{Specify any secondary language of the form. Form can have more than one, but this will only check one at a time. (e.g. "Francais") Entered text must match language specification in the XLSForm columns.}
}
\value{
List of data.frames with results of each "check" in tibbles.
}
\description{
This function imports two XLSForm files and currently conducts a number of checks to compare each of the "survey" sheets
}
\examples{
\dontrun{compare_forms(form1location,form2location,"Francais")}
}
