% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_quest.R
\name{export_quest}
\alias{export_quest}
\title{Export dataframe or flextable with readable version of xlsform questionnaire}
\usage{
export_quest(xlsformfile, primary = NULL, secondary = NULL, flex = TRUE)
}
\arguments{
\item{xlsformfile}{Filepath of xlsform as text or referring to object with filepath.}

\item{primary}{Main language defined in form. Specified as "label:English", etc.}

\item{secondary}{Specify any secondary language of the form. Form can have more than one, but this will only check one at a time. (e.g. "Francais") Entered text must match language specification in the XLSForm columns.}

\item{flex}{Specify whether flextable object is output (default) or tibble object}
}
\value{
Large flextable or tibble.
}
\description{
This function imports a XLSForm file and exports either a dataframe (tibble) or a flextable object with a readable version of questionnaire.
}
