% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_comparison.R
\name{tabulate_comparison}
\alias{tabulate_comparison}
\title{Tabulates results from checks output from "compare_survey"}
\usage{
tabulate_comparison(results, number)
}
\arguments{
\item{results}{List output by function 'compare_survey'.}

\item{number}{Number of check to tabulate.}
}
\value{
Flextable object.
}
\description{
This function creates flextable formatted tables from objects stored in full_compare list output by compare_survey.
}
\examples{
\dontrun{tabulate_comparison(1)}
}
